Improve AKA, add useful scripts
Previously AKA support was through duplicate code to Links. This update unifies the two.

Previous AKAs code and classes are removed
Links upgraded from "from,to" pairs to "subject, predicate, object" triples where the predicate might currently be "links-to" (previous behaviour and current default) or "aka-of" (subject is an AKA of object)

When viewing a Thing, connections are now prepended with a visual marker of the nature of that connection: '- ' is a link, '* ' is an aka. '!!' would indicate an unexpected kind of connection.

Two new useful scripts are added in manual_scripts/
removeCommonEnglish/remove.php
  Show those words and phrases in a source that are not in a dictionary, are not in the database, and are not otherwise excluded.
interesting.php
  Show those words and phrases (or the head if they are an aka) in a source that ARE in the database