Duplicate, nuance handling and minor clean ups
Allow behaviour in the face of duplicates to be passed down to the moment of adding the Thing to the database. So arrange that Manual additions allow duplicates, but other use - such as automated scripts - disallow it. This allows automated scripts to be idempotent, and the user to override that.

Show nuance as part of the Add Thing search

Show nuance as part of the show Thing screen

Accomodate that the nuance can be null or '' while unable to replicate bug

Accomodate that the or-words dictionary may be empty

Remove ';' menu item to deprecated functionality